Moved reusable ExteriorLevelData functions into utils files.
