+++ OpenTESArena/src/World/ExteriorLevelData.cpp
@@ -1,7 +1,9 @@
 #include <algorithm>
 #include <iomanip>
 
+#include "CityLevelUtils.h"
 #include "ExteriorLevelData.h"
+#include "WildLevelUtils.h"
 #include "WorldType.h"
 #include "../Assets/COLFile.h"
 #include "../Assets/MIFUtils.h"
@@ -25,116 +27,6 @@ ExteriorLevelData::ExteriorLevelData(SNInt gridWidth, int gridHeight, WEInt grid
 ExteriorLevelData::~ExteriorLevelData()
 {
 
-}
-
-void ExteriorLevelData::generateCity(uint32_t citySeed, int cityDim, WEInt gridDepth,
-	const std::vector<uint8_t> &reservedBlocks, const OriginalInt2 &startPosition,
-	ArenaRandom &random, const MiscAssets &miscAssets, Buffer2D<uint16_t> &dstFlor,
-	Buffer2D<uint16_t> &dstMap1, Buffer2D<uint16_t> &dstMap2)
-{
-	// Get the city's local X and Y, to be used later for building name generation.
-	const Int2 localCityPoint = LocationUtils::getLocalCityPoint(citySeed);
-
-	const int citySize = cityDim * cityDim;
-	std::vector<MIFUtils::BlockType> plan(citySize, MIFUtils::BlockType::Empty);
-
-	auto placeBlock = [citySize, &plan, &random](MIFUtils::BlockType blockType)
-	{
-		int planIndex;
-
-		do
-		{
-			planIndex = random.next() % citySize;
-		} while (plan.at(planIndex) != MIFUtils::BlockType::Empty);
-
-		plan.at(planIndex) = blockType;
-	};
-
-	// Set reserved blocks.
-	for (const uint8_t block : reservedBlocks)
-	{
-		// The original engine uses a fixed array so all block indices always fall within the
-		// plan, but since a dynamic array is used here, it has to ignore out-of-bounds blocks
-		// explicitly.
-		if (block < plan.size())
-		{
-			plan.at(block) = MIFUtils::BlockType::Reserved;
-		}
-	}
-
-	// Initial block placement.
-	placeBlock(MIFUtils::BlockType::Equipment);
-	placeBlock(MIFUtils::BlockType::MagesGuild);
-	placeBlock(MIFUtils::BlockType::NobleHouse);
-	placeBlock(MIFUtils::BlockType::Temple);
-	placeBlock(MIFUtils::BlockType::Tavern);
-	placeBlock(MIFUtils::BlockType::Spacer);
-
-	// Create city plan according to RNG.
-	const int emptyBlocksInPlan = static_cast<int>(
-		std::count(plan.begin(), plan.end(), MIFUtils::BlockType::Empty));
-	for (int remainingBlocks = emptyBlocksInPlan; remainingBlocks > 0; remainingBlocks--)
-	{
-		const MIFUtils::BlockType blockType = MIFUtils::generateRandomBlockType(random);
-		placeBlock(blockType);
-	}
-
-	// Build the city, loading data for each block. Load blocks right to left, top to bottom.
-	WEInt xDim = 0;
-	SNInt zDim = 0;
-
-	for (const MIFUtils::BlockType block : plan)
-	{
-		if (block != MIFUtils::BlockType::Reserved)
-		{
-			const std::string blockMifName = MIFUtils::makeCityBlockMifName(block, random);
-
-			// Load the block's .MIF data into the level.
-			const auto &cityBlockMifs = miscAssets.getCityBlockMifs();
-			const auto iter = cityBlockMifs.find(blockMifName);
-			if (iter == cityBlockMifs.end())
-			{
-				DebugCrash("Could not find .MIF file \"" + blockMifName + "\".");
-			}
-
-			const MIFFile &blockMif = iter->second;
-			const WEInt blockWidth = blockMif.getWidth();
-			const SNInt blockDepth = blockMif.getDepth();
-			const auto &blockLevel = blockMif.getLevels().front();
-			const BufferView2D<const MIFFile::VoxelID> blockFLOR = blockLevel.getFLOR();
-			const BufferView2D<const MIFFile::VoxelID> blockMAP1 = blockLevel.getMAP1();
-			const BufferView2D<const MIFFile::VoxelID> blockMAP2 = blockLevel.getMAP2();
-
-			// Offset of the block in the voxel grid.
-			const WEInt xOffset = startPosition.x + (xDim * 20);
-			const SNInt zOffset = startPosition.y + (zDim * 20);
-
-			// Copy block data to temp buffers.
-			for (SNInt z = 0; z < blockDepth; z++)
-			{
-				for (WEInt x = 0; x < blockWidth; x++)
-				{
-					const MIFFile::VoxelID srcFlorVoxel = blockFLOR.get(x, z);
-					const MIFFile::VoxelID srcMap1Voxel = blockMAP1.get(x, z);
-					const MIFFile::VoxelID srcMap2Voxel = blockMAP2.get(x, z);
-					const WEInt dstX = xOffset + x;
-					const SNInt dstZ = zOffset + z;
-					dstFlor.set(dstX, dstZ, srcFlorVoxel);
-					dstMap1.set(dstX, dstZ, srcMap1Voxel);
-					dstMap2.set(dstX, dstZ, srcMap2Voxel);
-				}
-			}
-		}
-
-		xDim++;
-
-		// Move to the next row if done with the current one.
-		if (xDim == cityDim)
-		{
-			xDim = 0;
-			zDim++;
-		}
-	}
 }
 
 void ExteriorLevelData::generateBuildingNames(const LocationDefinition &locationDef,
@@ -472,417 +364,6 @@ void ExteriorLevelData::generateWildChunkBuildingNames(const ExeData &exeData)
 	}
 }
 
-void ExteriorLevelData::revisePalaceGraphics(Buffer2D<uint16_t> &map1, SNInt gridWidth, WEInt gridDepth)
-{
-	// Lambda for obtaining a two-byte MAP1 voxel.
-	auto getMap1Voxel = [&map1, gridWidth, gridDepth](SNInt x, WEInt z)
-	{
-		const uint16_t voxel = map1.get(z, x);
-		return voxel;
-	};
-
-	auto setMap1Voxel = [&map1, gridWidth, gridDepth](SNInt x, WEInt z, uint16_t voxel)
-	{
-		map1.set(z, x, voxel);
-	};
-
-	struct SearchResult
-	{
-		enum class Side { None, North, South, East, West };
-
-		Side side;
-
-		// Distance from the associated origin dimension, where (0, 0) is at the top right.
-		int offset;
-
-		SearchResult(Side side, int offset)
-		{
-			this->side = side;
-			this->offset = offset;
-		}
-	};
-
-	// Find one of the palace graphic blocks, then extrapolate the positions of
-	// the other palace graphic and the gates.
-	const SearchResult result = [gridWidth, gridDepth, &getMap1Voxel]()
-	{
-		auto isPalaceBlock = [&getMap1Voxel](SNInt x, WEInt z)
-		{
-			const uint16_t voxel = getMap1Voxel(x, z);
-			const uint8_t mostSigNibble = (voxel & 0xF000) >> 12;
-			return mostSigNibble == 0x9;
-		};
-
-		// North (top edge) and south (bottom edge), search right to left.
-		for (WEInt z = 1; z < (gridDepth - 1); z++)
-		{
-			const SNInt northX = 0;
-			const SNInt southX = gridWidth - 1;
-			if (isPalaceBlock(northX, z))
-			{
-				return SearchResult(SearchResult::Side::North, z);
-			}
-			else if (isPalaceBlock(southX, z))
-			{
-				return SearchResult(SearchResult::Side::South, z);
-			}
-		}
-
-		// East (right edge) and west (left edge), search top to bottom.
-		for (SNInt x = 1; x < (gridWidth - 1); x++)
-		{
-			const WEInt eastZ = 0;
-			const WEInt westZ = gridDepth - 1;
-			if (isPalaceBlock(x, eastZ))
-			{
-				return SearchResult(SearchResult::Side::East, x);
-			}
-			else if (isPalaceBlock(x, westZ))
-			{
-				return SearchResult(SearchResult::Side::West, x);
-			}
-		}
-
-		// No palace gate found. This should never happen because every city/town/village
-		// in the original game has a palace gate somewhere.
-		return SearchResult(SearchResult::Side::None, 0);
-	}();
-
-	// Decide how to extrapolate the search results.
-	if (result.side != SearchResult::Side::None)
-	{
-		// The direction to step from a palace voxel to the other palace voxel.
-		const NewInt2 northSouthPalaceStep = VoxelUtils::West;
-		const NewInt2 eastWestPalaceStep = VoxelUtils::South;
-
-		// Gets the distance in voxels from a palace voxel to its gate, or -1 if no gate exists.
-		const int NO_GATE = -1;
-		auto getGateDistance = [&getMap1Voxel, NO_GATE](const NewInt2 &palaceVoxel, const NewInt2 &dir)
-		{
-			auto isGateBlock = [&getMap1Voxel](SNInt x, WEInt z)
-			{
-				const uint16_t voxel = getMap1Voxel(x, z);
-				const uint8_t mostSigNibble = (voxel & 0xF000) >> 12;
-				return mostSigNibble == 0xA;
-			};
-
-			// Gates should usually be within a couple blocks of their castle graphic. If not,
-			// then no gate exists.
-			const int MAX_GATE_DIST = 8;
-
-			int i = 0;
-			NewInt2 position = palaceVoxel;
-			while ((i < MAX_GATE_DIST) && !isGateBlock(position.x, position.y))
-			{
-				position = position + dir;
-				i++;
-			}
-
-			return (i < MAX_GATE_DIST) ? i : NO_GATE;
-		};
-
-		// Set the positions of the two palace voxels and the two gate voxels.
-		NewInt2 firstPalaceVoxel, secondPalaceVoxel, firstGateVoxel, secondGateVoxel;
-		uint16_t firstPalaceVoxelID, secondPalaceVoxelID, gateVoxelID;
-		int gateDist;
-		if (result.side == SearchResult::Side::North)
-		{
-			firstPalaceVoxel = NewInt2(0, result.offset);
-			secondPalaceVoxel = firstPalaceVoxel + northSouthPalaceStep;
-			const NewInt2 gateDir = VoxelUtils::South;
-			gateDist = getGateDistance(firstPalaceVoxel, gateDir);
-			firstGateVoxel = firstPalaceVoxel + (gateDir * gateDist);
-			secondGateVoxel = firstGateVoxel + northSouthPalaceStep;
-			firstPalaceVoxelID = 0xA5B4;
-			secondPalaceVoxelID = 0xA5B5;
-			gateVoxelID = 0xA1B3;
-		}
-		else if (result.side == SearchResult::Side::South)
-		{
-			firstPalaceVoxel = NewInt2(gridWidth - 1, result.offset);
-			secondPalaceVoxel = firstPalaceVoxel + northSouthPalaceStep;
-			const NewInt2 gateDir = VoxelUtils::North;
-			gateDist = getGateDistance(firstPalaceVoxel, gateDir);
-			firstGateVoxel = firstPalaceVoxel + (gateDir * gateDist);
-			secondGateVoxel = firstGateVoxel + northSouthPalaceStep;
-			firstPalaceVoxelID = 0xA535;
-			secondPalaceVoxelID = 0xA534;
-			gateVoxelID = 0xA133;
-		}
-		else if (result.side == SearchResult::Side::East)
-		{
-			firstPalaceVoxel = NewInt2(result.offset, 0);
-			secondPalaceVoxel = firstPalaceVoxel + eastWestPalaceStep;
-			const NewInt2 gateDir = VoxelUtils::West;
-			gateDist = getGateDistance(firstPalaceVoxel, gateDir);
-			firstGateVoxel = firstPalaceVoxel + (gateDir * gateDist);
-			secondGateVoxel = firstGateVoxel + eastWestPalaceStep;
-			firstPalaceVoxelID = 0xA575;
-			secondPalaceVoxelID = 0xA574;
-			gateVoxelID = 0xA173;
-		}
-		else if (result.side == SearchResult::Side::West)
-		{
-			firstPalaceVoxel = NewInt2(result.offset, gridDepth - 1);
-			secondPalaceVoxel = firstPalaceVoxel + eastWestPalaceStep;
-			const NewInt2 gateDir = VoxelUtils::East;
-			gateDist = getGateDistance(firstPalaceVoxel, gateDir);
-			firstGateVoxel = firstPalaceVoxel + (gateDir * gateDist);
-			secondGateVoxel = firstGateVoxel + eastWestPalaceStep;
-			firstPalaceVoxelID = 0xA5F4;
-			secondPalaceVoxelID = 0xA5F5;
-			gateVoxelID = 0xA1F3;
-		}
-
-		// Set the voxel IDs to their new values.
-		setMap1Voxel(firstPalaceVoxel.x, firstPalaceVoxel.y, firstPalaceVoxelID);
-		setMap1Voxel(secondPalaceVoxel.x, secondPalaceVoxel.y, secondPalaceVoxelID);
-
-		if (gateDist != NO_GATE)
-		{
-			setMap1Voxel(firstGateVoxel.x, firstGateVoxel.y, gateVoxelID);
-			setMap1Voxel(secondGateVoxel.x, secondGateVoxel.y, gateVoxelID);
-		}
-	}
-	else
-	{
-		// The search did not find any palace graphics block.
-		DebugLogWarning("No palace graphics found to revise.");
-	}
-}
-
-Buffer2D<ExteriorLevelData::WildBlockID> ExteriorLevelData::generateWildernessIndices(
-	uint32_t wildSeed, const ExeData::Wilderness &wildData)
-{
-	constexpr int wildWidth = 64;
-	constexpr int wildHeight = 64;
-	Buffer2D<WildBlockID> indices(wildWidth, wildHeight);
-	ArenaRandom random(wildSeed);
-
-	// Generate a random wilderness .MIF index for each wilderness chunk.
-	std::generate(indices.get(), indices.get() + (indices.getWidth() * indices.getHeight()),
-		[&wildData, &random]()
-	{
-		// Determine the wilderness block list to draw from.
-		const auto &blockList = [&wildData, &random]() -> const std::vector<uint8_t>&
-		{
-			const uint16_t normalVal = 0x6666;
-			const uint16_t villageVal = 0x4000;
-			const uint16_t dungeonVal = 0x2666;
-			const uint16_t tavernVal = 0x1999;
-			int randVal = random.next();
-
-			if (randVal < normalVal)
-			{
-				return wildData.normalBlocks;
-			}
-			else
-			{
-				randVal -= normalVal;
-				if (randVal < villageVal)
-				{
-					return wildData.villageBlocks;
-				}
-				else
-				{
-					randVal -= villageVal;
-					if (randVal < dungeonVal)
-					{
-						return wildData.dungeonBlocks;
-					}
-					else
-					{
-						randVal -= dungeonVal;
-						if (randVal < tavernVal)
-						{
-							return wildData.tavernBlocks;
-						}
-						else
-						{
-							return wildData.templeBlocks;
-						}
-					}
-				}
-			}
-		}();
-
-		const int blockListIndex = (random.next() & 0xFF) % blockList.size();
-		return blockList[blockListIndex];
-	});
-
-	// City indices in the center of the wilderness (WILD001.MIF, etc.).
-	DebugAssertMsg(wildWidth >= 2, "Wild width \"" + std::to_string(wildWidth) + "\" too small.");
-	DebugAssertMsg(wildHeight >= 2, "Wild height \"" + std::to_string(wildHeight) + "\" too small.");
-	const int cityX = (wildWidth / 2) - 1;
-	const int cityY = (wildHeight / 2) - 1;
-	indices.set(cityX, cityY, 1);
-	indices.set(cityX + 1, cityY, 2);
-	indices.set(cityX, cityY + 1, 3);
-	indices.set(cityX + 1, cityY + 1, 4);
-
-	return indices;
-}
-
-void ExteriorLevelData::reviseWildernessCity(const LocationDefinition &locationDef,
-	Buffer2D<uint16_t> &flor, Buffer2D<uint16_t> &map1, Buffer2D<uint16_t> &map2,
-	const MiscAssets &miscAssets)
-{
-	// For now, assume the given buffers are for the entire 4096x4096 wilderness.
-	// @todo: change to only care about 128x128 layers.
-	DebugAssert(flor.getWidth() == (64 * RMDFile::WIDTH));
-	DebugAssert(flor.getWidth() == flor.getHeight());
-	DebugAssert(flor.getWidth() == map1.getWidth());
-	DebugAssert(flor.getWidth() == map2.getWidth());
-
-	// Clear all placeholder city blocks.
-	const int placeholderWidth = RMDFile::WIDTH * 2;
-	const int placeholderDepth = RMDFile::DEPTH * 2;
-
-	// @todo: change to only care about 128x128 floors -- these should both be removed.
-	const WEInt xOffset = RMDFile::WIDTH * 31;
-	const SNInt zOffset = RMDFile::DEPTH * 31;
-
-	for (WEInt x = 0; x < placeholderWidth; x++)
-	{
-		const int startIndex = zOffset + ((x + xOffset) * flor.getHeight());
-
-		auto clearRow = [placeholderDepth, startIndex](Buffer2D<uint16_t> &dst)
-		{
-			const auto dstBegin = dst.get() + startIndex;
-			const auto dstEnd = dstBegin + placeholderDepth;
-			DebugAssert(dstEnd <= dst.end());
-			std::fill(dstBegin, dstEnd, 0);
-		};
-
-		clearRow(flor);
-		clearRow(map1);
-		clearRow(map2);
-	}
-
-	// Get city generation info -- the .MIF filename to load for the city skeleton.
-	const LocationDefinition::CityDefinition &cityDef = locationDef.getCityDefinition();
-	const std::string mifName = cityDef.mapFilename;
-	MIFFile mif;
-	if (!mif.init(mifName.c_str()))
-	{
-		DebugLogError("Couldn't init .MIF file \"" + mifName + "\".");
-		return;
-	}
-
-	const MIFFile::Level &level = mif.getLevels().front();
-	const BufferView2D<const MIFFile::VoxelID> levelFLOR = level.getFLOR();
-	const BufferView2D<const MIFFile::VoxelID> levelMAP1 = level.getMAP1();
-	const BufferView2D<const MIFFile::VoxelID> levelMAP2 = level.getMAP2();
-
-	// Buffers for the city data. Copy the .MIF data into them.
-	Buffer2D<uint16_t> cityFlor(mif.getWidth(), mif.getDepth());
-	Buffer2D<uint16_t> cityMap1(mif.getWidth(), mif.getDepth());
-	Buffer2D<uint16_t> cityMap2(mif.getWidth(), mif.getDepth());
-
-	for (WEInt x = 0; x < mif.getWidth(); x++)
-	{
-		for (SNInt z = 0; z < mif.getDepth(); z++)
-		{
-			const MIFFile::VoxelID srcFlorVoxel = levelFLOR.get(x, z);
-			const MIFFile::VoxelID srcMap1Voxel = levelMAP1.get(x, z);
-			const MIFFile::VoxelID srcMap2Voxel = levelMAP2.get(x, z);
-			cityFlor.set(x, z, srcFlorVoxel);
-			cityMap1.set(x, z, srcMap1Voxel);
-			cityMap2.set(x, z, srcMap2Voxel);
-		}
-	}
-
-	// Run city generation if it's not a premade city. The center province's city does not have
-	// any special generation -- the .MIF buffers are simply used as-is (with some simple palace
-	// gate revisions done afterwards).
-	const bool isPremadeCity = cityDef.premade;
-	if (!isPremadeCity)
-	{
-		const int cityBlocksPerSide = cityDef.cityBlocksPerSide;
-		const std::vector<uint8_t> &reservedBlocks = *cityDef.reservedBlocks;
-		const OriginalInt2 blockStartPosition(cityDef.blockStartPosX, cityDef.blockStartPosY);
-		const uint32_t citySeed = cityDef.citySeed;
-		ArenaRandom random(citySeed);
-
-		// Write generated city data into the temp city buffers.
-		ExteriorLevelData::generateCity(citySeed, cityBlocksPerSide, mif.getWidth(),
-			reservedBlocks, blockStartPosition, random, miscAssets, cityFlor, cityMap1, cityMap2);
-	}
-
-	// Transform city voxels based on the wilderness rules.
-	for (WEInt x = 0; x < mif.getWidth(); x++)
-	{
-		for (SNInt z = 0; z < mif.getDepth(); z++)
-		{
-			uint16_t &map1Voxel = cityMap1.get(x, z);
-			uint16_t &map2Voxel = cityMap2.get(x, z);
-
-			if ((map1Voxel & 0x8000) != 0)
-			{
-				map1Voxel = 0;
-				map2Voxel = 0;
-			}
-			else
-			{
-				const bool isWall = (map1Voxel == 0x2F2F) || (map1Voxel == 0x2D2D) || (map1Voxel == 0x2E2E);
-				if (!isWall)
-				{
-					map1Voxel = 0;
-					map2Voxel = 0;
-				}
-				else
-				{
-					// Replace solid walls.
-					if (map1Voxel == 0x2F2F)
-					{
-						map1Voxel = 0x3030;
-						map2Voxel = 0x3030 | (map2Voxel & 0x8080);
-					}
-					else if (map1Voxel == 0x2D2D)
-					{
-						map1Voxel = 0x2F2F;
-						map2Voxel = 0x3030 | (map2Voxel & 0x8080);
-					}
-					else if (map1Voxel == 0x2E2E)
-					{
-						map2Voxel = 0x3030 | (map2Voxel & 0x8080);
-					}
-				}
-			}
-		}
-	}
-
-	// Write city buffers into the wilderness.
-	for (SNInt z = 0; z < mif.getDepth(); z++)
-	{
-		for (WEInt x = 0; x < mif.getWidth(); x++)
-		{
-			const uint16_t srcFlorVoxel = cityFlor.get(x, z);
-			const uint16_t srcMap1Voxel = cityMap1.get(x, z);
-			const uint16_t srcMap2Voxel = cityMap2.get(x, z);
-			const WEInt dstX = xOffset + x;
-			const SNInt dstZ = zOffset + z;
-			flor.set(dstX, dstZ, srcFlorVoxel);
-			map1.set(dstX, dstZ, srcMap1Voxel);
-			map2.set(dstX, dstZ, srcMap2Voxel);
-		}
-	}
-}
-
-OriginalInt2 ExteriorLevelData::getRelativeWildOrigin(const Int2 &voxel)
-{
-	return OriginalInt2(
-		voxel.x - (voxel.x % (RMDFile::WIDTH * 2)),
-		voxel.y - (voxel.y % (RMDFile::DEPTH * 2)));
-}
-
-NewInt2 ExteriorLevelData::getCenteredWildOrigin(const NewInt2 &voxel)
-{
-	return NewInt2(
-		(std::max(voxel.x - 32, 0) / RMDFile::WIDTH) * RMDFile::WIDTH,
-		(std::max(voxel.y - 32, 0) / RMDFile::DEPTH) * RMDFile::DEPTH);
-}
-
 ExteriorLevelData ExteriorLevelData::loadCity(const LocationDefinition &locationDef,
 	const ProvinceDefinition &provinceDef, const MIFFile::Level &level, WeatherType weatherType,
 	int currentDay, int starCount, const std::string &infName, SNInt gridWidth, WEInt gridDepth,
@@ -924,12 +405,12 @@ ExteriorLevelData ExteriorLevelData::loadCity(const LocationDefinition &location
 		// Generate procedural city data and write it into the temp buffers.
 		const std::vector<uint8_t> &reservedBlocks = *cityDef.reservedBlocks;
 		const OriginalInt2 blockStartPosition(cityDef.blockStartPosX, cityDef.blockStartPosY);
-		ExteriorLevelData::generateCity(citySeed, cityDef.cityBlocksPerSide, gridDepth,
+		CityLevelUtils::generateCity(citySeed, cityDef.cityBlocksPerSide, gridDepth,
 			reservedBlocks, blockStartPosition, random, miscAssets, tempFlor, tempMap1, tempMap2);
 	}
 
 	// Run the palace gate graphic algorithm over the perimeter of the MAP1 data.
-	ExteriorLevelData::revisePalaceGraphics(tempMap1, gridWidth, gridDepth);
+	CityLevelUtils::revisePalaceGraphics(tempMap1, gridWidth, gridDepth);
 
 	// Create the level for the voxel data to be written into.
 	ExteriorLevelData levelData(gridWidth, level.getHeight(), gridDepth, infName, level.getName());
@@ -966,9 +447,9 @@ ExteriorLevelData ExteriorLevelData::loadWilderness(const LocationDefinition &lo
 	TextureManager &textureManager)
 {
 	const LocationDefinition::CityDefinition &cityDef = locationDef.getCityDefinition();
-	const auto &wildData = miscAssets.getExeData().wild;
+	const ExeData::Wilderness &wildData = miscAssets.getExeData().wild;
 	const Buffer2D<WildBlockID> wildIndices =
-		ExteriorLevelData::generateWildernessIndices(cityDef.wildSeed, wildData);
+		WildLevelUtils::generateWildernessIndices(cityDef.wildSeed, wildData);
 
 	// Temp buffers for voxel data.
 	Buffer2D<uint16_t> tempFlor(RMDFile::DEPTH * wildIndices.getWidth(),
@@ -1018,7 +499,7 @@ ExteriorLevelData ExteriorLevelData::loadWilderness(const LocationDefinition &lo
 	}
 
 	// Change the placeholder WILD00{1..4}.MIF blocks to the ones for the given city.
-	ExteriorLevelData::reviseWildernessCity(locationDef, tempFlor, tempMap1, tempMap2, miscAssets);
+	WildLevelUtils::reviseWildernessCity(locationDef, tempFlor, tempMap1, tempMap2, miscAssets);
 
 	// Create the level for the voxel data to be written into.
 	const int levelHeight = 6;
